pyproject.toml
[project]
requires-python = ">=3.10"
name = "denots"
version = "0.0.0"
dependencies = [
    "aeon==1.0.0",
    "causal-conv1d>=1.5.0.post8",
    "funcy>=2.0",
    "gmpy2>=2.2.1",
    "great-tables[extra]>=0.17.0",
    "hydra-core>=1.3.2",
    "imgkit>=1.2.3",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "kaggle>=1.6.17",
    "loguru>=0.7.3",
    "mamba-ssm[causal-conv1d]>=2.2.4",
    "matplotlib>=3.10.0",
    "mlflow>=2.20.1",
    "mlflow-export-import",
    "mlflow-skinny>=2.20.2",
    "pandas<2.1.0",
    "physiopro",
    "polars==1.21.0",
    "psutil>=7.0.0",
    "pyarrow>=18.1.0",
    "pynvml>=12.0.0",
    "pytorch-lightning>=2.5.0.post0",
    "ruff>=0.15.1",
    "scikit-learn>=1.0.0,<1.6.0",
    "selenium>=4.30.0",
    "signatory==1.2.6.1.9.0",
    "tensordict==0.6.2",
    "torch==2.6.0",
    "torchao==0.6.1",
    "torchcde>=0.2.5",
    "torchinfo>=1.8.0",
    "torchmetrics>=1.6.1",
    "torchode>=1.0.0",
    "torchtune==0.3.1",
    "tqdm>=4.67.1",
    "xlsxwriter>=3.2.2",
]

[tool.uv]
override-dependencies=["torch==2.6.0"]

[tool.uv.extra-build-dependencies]
causal-conv1d = ["torch"]

[tool.uv.sources]
physiopro = { git = "https://github.com/microsoft/PhysioPro.git" }
mlflow-export-import = { git = "https://github.com/mlflow/mlflow-export-import/" }

[tool.ruff.lint]
extend-select = [
  "UP",  # pyupgrade
  "D",   # pydocstyle
  "I",   # isort
  "W"    # pycodestyle
]

ignore = [
  "W191",
  "E111",
  "E114",
  "E117",
  "D206",
  "D300",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "COM812",
  "COM819",
  "ISC001",
  "ISC002"
]

[tool.ruff.lint.pycodestyle]
max-doc-length = 120

[dependency-groups]
dev = [
    "sympy>=1.13.1",
]

